+++ src/decorators/createMethodDecorator.ts
@@ -1,7 +1,7 @@
 import type { MiddlewareFn } from "@/typings/Middleware";
 import { UseMiddleware } from "./UseMiddleware";
 
-export function createMethodDecorator<TContextType = {}>(
+export function createMethodDecorator<TContextType extends object = object>(
   resolver: MiddlewareFn<TContextType>,
 ): MethodDecorator {
   return UseMiddleware(resolver);