+++ src/decorators/createParamDecorator.ts
@@ -2,7 +2,7 @@ import { SymbolKeysNotSupportedError } from "@/errors";
 import { getMetadataStorage } from "@/metadata/getMetadataStorage";
 import type { ParameterDecorator, ResolverData } from "@/typings";
 
-export function createParamDecorator<TContextType = {}>(
+export function createParamDecorator<TContextType extends object = object>(
   resolver: (resolverData: ResolverData<TContextType>) => any,
 ): ParameterDecorator {
   return (prototype, propertyKey, parameterIndex) => {