+++ CHANGELOG.md
@@ -8,9 +8,15 @@
 
 - **Breaking Change**: upgrade `ArgumentValidationError` and replace `UnauthorizedError` and `ForbiddenError` with `AuthenticationError`, `AuthorizationError` that are extending `GraphQLError` to let the error details be accessible in the `extensions` property
 - **Breaking Change**: change `ClassType` constraint from `ClassType<T = any>` to `ClassType<T extends object = object>` in order to make it work properly with new TS features
-- **Breaking Change**: removed `dateScalarMode` option from `buildSchema`
+- **Breaking Change**: remove `dateScalarMode` option from `buildSchema`
 - **Breaking Change**: make `graphql-scalars` package a peer dependency and use date scalars from it instead of custom ones
 - **Breaking Change**: exported `GraphQLISODateTime` scalar has now a name `DateTimeISO`
+- **Breaking Change**: change `ValidatorFn` signature from `ValidatorFn<TArgs>` to `ValidatorFn<TContext>`
+- support custom validation function getting resolver data on validate
+
+### Fixes
+
+- allow `ValidatorFn` to accept array of values (instead of only `object | undefined`)
 
 ## v2.0.0-beta.2